Add in some of the libressl/openssh compatibility code.
Out of curiosity, try to get a feel for how hard this would be to port
to Windows/PDcurses.  If it could be made to work it would be a more
powerful solution than yedit.exe for Windows fallback, but the port may
not be trivial.  openssh/libressl need some of what this code uses, so
stage those files, but a compile test proves there's a lot more to do.